requirements.docker.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
PyYAML>=6.0.1

ultralytics>=8.3.0
onnx>=1.16.0
onnxruntime>=1.18.0
onnxslim>=0.1.71

numpy>=1.26.0
opencv-contrib-python-headless>=4.10.0.84
scikit-image>=0.24.0
pandas>=2.2.2
matplotlib>=3.9.0
reportlab>=4.2.2
Pillow>=10.4.0

pytest>=8.3.0
pytest-asyncio>=1.3.0
httpx>=0.27.0
tensorboard>=2.17.0

datasets>=2.20.0
huggingface_hub>=0.24.0
pycocotools>=2.0.8
nibabel>=5.2.1
ijson>=3.3.0
SQLAlchemy>=2.0.36
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0
redis>=5.2.0
rq==1.16.2
prometheus-fastapi-instrumentator>=7.0.0
albumentations>=1.4.18
wandb>=0.18.7
